deeper method clone w/ subname for mro
mro uses the internal subname, see the mro docs /"defined outside".

With the sub installed in the anon class but still internally tagged
with the name of the originating class, I got this error:

  No next::method 'set' found for Class::MOP::Class::__ANON__..

The subname is settable with Sub::Name and gettable with Sub::Identify

The method cloning is clunky, but Class::MOP::Method didn't seem to
allow you to set some attributes or provide a "new" (although new_object
might do).  The test passes.

Sub::Clone is used so that the subname doesn't effect the original
HashStore::get, etc.